ExploringLLMs/GoogleLLMs/asr_google/requirements.txt
SpeechRecognition
PyAudio
google-genai
python-dotenv